tool_executor.py
from dotenv import load_dotenv
from langchain_tavily import TavilySearch
from langchain_core.tools import StructuredTool
from langgraph.prebuilt import ToolNode
from schemas import AnswerQuestion, ReviseAnswer

#this will be the node that will execute the tools



tavily_tool = TavilySearch(max_results=5)


#going to have two tavily tools
#each tool serves a different purpose in the application workflow
    # answer questiont tool
    # revise answer tool 
#both tools will execute tavily search.

def run_queries(search_queries: list[str], **kwargs):
    """Run the generated queries"""
    return tavily_tool.batch([{"query": query} for query in search_queries])


execute_tools = ToolNode(
    [
        StructuredTool.from_function(run_queries, name=AnswerQuestion.__name__),
        StructuredTool.from_function(run_queries, name=ReviseAnswer.__name__)
    ]
)